adding homework 9 chapter 10 (read description)
Adding all I am going to do for 10-10 from the book's own questions. The
book is referencing non-existent things to look at to do the assignment
and after at least one issue an assignment two questions in a row in
this chapter of this crap, I am finally deeming the book as being
incompetent when it comes to the questions.